refactor(#53): 모니터링 큐 화면 구조 분리
