+++ src/features/monitoring-queue/model/types.ts
@@ -1,0 +1,11 @@
+export type QueueStatusItem = {
+  type: "image" | "video";
+  model: string;
+  pending: number;
+  processing: number;
+};
+
+export type QueueStatusResponse = {
+  updatedAt: string;
+  items: QueueStatusItem[];
+};
